+++ src/main/java/team7/inplace/video/application/VideoService.java
@@ -4,8 +4,8 @@ import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import team7.inplace.influencer.entity.Influencer;
-import team7.inplace.influencer.repository.InfluencerRepository;
+import team7.inplace.influencer.domain.Influencer;
+import team7.inplace.influencer.persistence.InfluencerRepository;
 import team7.inplace.place.application.dto.PlaceForVideo;
 import team7.inplace.place.domain.Place;
 import team7.inplace.video.application.dto.VideoInfo;
