+++ src/main/java/team7/inplace/video/domain/Video.java
@@ -1,11 +1,17 @@
 package team7.inplace.video.domain;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
-import team7.inplace.influencer.entity.Influencer;
+import team7.inplace.influencer.domain.Influencer;
 import team7.inplace.place.domain.Place;
 
 @Entity
